refactor cli.py
