+++ mcp_fuzzer/cli/__init__.py
@@ -1,0 +1,32 @@
+from .main import run_cli
+from .args import (
+    build_unified_client_args,
+    create_argument_parser,
+    get_cli_config,
+    parse_arguments,
+    print_startup_info,
+    setup_logging,
+    validate_arguments,
+)
+from ..auth import load_auth_config, setup_auth_from_env
+from ..transport import create_transport
+import asyncio  # re-export for tests that patch mcp_fuzzer.cli.asyncio.run
+from rich.console import Console
+import sys  # re-export for tests that patch mcp_fuzzer.cli.sys.exit
+
+__all__ = [
+    "run_cli",
+    "build_unified_client_args",
+    "create_argument_parser",
+    "get_cli_config",
+    "parse_arguments",
+    "print_startup_info",
+    "setup_logging",
+    "validate_arguments",
+    "load_auth_config",
+    "setup_auth_from_env",
+    "create_transport",
+    "asyncio",
+    "Console",
+    "sys",
+]
